rename encoder and format class names
